rename HeroHub.jsx to Hero.jsx
